added user verificaton on the client side, and refactored graph view

+++ server/routes/stopwatches.js
@@ -6,7 +6,7 @@ const db = require("../database");
 
 router.use((req, res, next) => {
   if (!req.session) {
-    return res.status(401).json({ error: "Unauthorized" });
+    return res.status(401).json({ error: "Unauthorized no session" });
   }
   db.get(`SELECT * FROM users WHERE id = ?`, [req.session.id], (err, row) => {
     if (err) {
@@ -14,16 +14,16 @@ router.use((req, res, next) => {
       return res.status(500).json({ error: "Error getting user" });
     }
     if (!row) {
-      return res.status(401).json({ error: "Unauthorized" });
+      return res.status(401).json({ error: "Unauthorized no user of this id" });
     }
 
     const inMemoryBuffer = Buffer.from(req.session.hashed_password, "utf-8");
     const dbBuffer = Buffer.from(row.hashed_password);
 
-    console.log(inMemoryBuffer);
-    console.log(dbBuffer);
     if (Buffer.compare(inMemoryBuffer, dbBuffer) !== 0) {
-      return res.status(401).json({ error: "Unauthorized" });
+      return res
+        .status(401)
+        .json({ error: "Unauthorized password does not match" });
     }
     next();
   });
@@ -44,7 +44,7 @@ router.get("/", (req, res) => {
 // Get a stopwatch of provided id for the logged-in user
 router.get("/:id", (req, res) => {
   const sql = `SELECT id, name, description FROM stopwatches WHERE id = ? AND user_id = ?`;
-  db.get(sql, [req.params.id, req.session.user.id], (err, row) => {
+  db.get(sql, [req.params.id, req.session.id], (err, row) => {
     if (err) {
       console.error(`Error getting stopwatches:`, err);
       return res.status(500).json({ error: "Error getting a stopwatch" });
@@ -56,11 +56,7 @@ router.get("/:id", (req, res) => {
 // Add a new stopwatch for the logged-in user
 router.post("/", (req, res) => {
   const sql = `INSERT INTO stopwatches(user_id, name, description) VALUES(?, ?, ?)`;
-  const values = [
-    req.session.user.id,
-    req.body.name,
-    req.body.description || "",
-  ];
+  const values = [req.session.id, req.body.name, req.body.description || ""];
   db.run(sql, values, function (err) {
     if (err) {
       console.error(`Error adding stopwatch:`, err);
@@ -77,8 +73,8 @@ router.post("/", (req, res) => {
 // Rename a stopwatch of provided id
 router.put("/:id", (req, res) => {
   const data = req.body;
-  const sql = `UPDATE stopwatches SET name = ? WHERE id = ?`;
-  const values = [data.name, req.params.id];
+  const sql = `UPDATE stopwatches SET name = ? WHERE id = ? AND user_id = ?`;
+  const values = [data.name, req.params.id, req.session.id];
   db.run(sql, values, function (err) {
     if (err) {
       console.error(`Error updating stopwatch ${req.params.id}:`, err);
@@ -95,8 +91,8 @@ router.put("/:id", (req, res) => {
 
 // Delete a stopwatch of provided id and all its entries
 router.delete("/:id", (req, res) => {
-  const sql2 = `DELETE FROM stopwatches_entries WHERE stopwatch_id = ?`;
-  db.run(sql2, [req.params.id], function (err) {
+  const sql2 = `DELETE FROM stopwatches_entries WHERE stopwatch_id = ? AND user_id = ?`;
+  db.run(sql2, [req.params.id, req.session.id], function (err) {
     if (err) {
       console.error(
         `Error deleting entries for stopwatch ${req.params.id}:`,
@@ -108,8 +104,8 @@ router.delete("/:id", (req, res) => {
     }
     console.log(`Row(s) deleted: ${this.changes}`);
 
-    const sql = `DELETE FROM stopwatches WHERE id = ?`;
-    db.run(sql, [req.params.id], function (err) {
+    const sql = `DELETE FROM stopwatches WHERE id = ? AND user_id = ?`;
+    db.run(sql, [req.params.id, req.session.id], function (err) {
       if (err) {
         console.error(`Error deleting stopwatch ${req.params.id}:`, err);
         return res
@@ -125,8 +121,8 @@ router.delete("/:id", (req, res) => {
 // Edit description for a stopwatch of provided id
 router.put("/:id/description", (req, res) => {
   const data = req.body;
-  const sql = `UPDATE stopwatches SET description = ? WHERE id = ?`;
-  const values = [data.description, req.params.id];
+  const sql = `UPDATE stopwatches SET description = ? WHERE id = ? AND user_id = ?`;
+  const values = [data.description, req.params.id, req.session.id];
   db.run(sql, values, function (err) {
     if (err) {
       console.error(
