+++ server/routes/stopwatches_entries.js
@@ -4,9 +4,9 @@ const router = express.Router();
 const db = require("../database");
 
 // Get 50 entries for a stopwatch of provided id
-router.get("/ :id/entries", (req, res) => {
-  const sql = `SELECT id, start_time, stop_time, note FROM stopwatches_entries WHERE stopwatch_id = ?  ORDER BY id DESC LIMIT 50`;
-  db.all(sql, [req.params.id, req.params.n], (err, rows) => {
+router.get("/:id/entries", (req, res) => {
+  const sql = `SELECT id, start_time, stop_time, note FROM stopwatches_entries WHERE stopwatch_id = ? AND user_id = ? ORDER BY id DESC LIMIT 50`;
+  db.all(sql, [req.params.id, req.session.id], (err, rows) => {
     if (err) {
       console.error(
         `Error getting entries for stopwatch ${req.params.id}:`,
@@ -22,8 +22,8 @@ router.get("/ :id/entries", (req, res) => {
 
 // Get n entries for a stopwatch of provided id
 router.get("/:id/entries/:n", (req, res) => {
-  const sql = `SELECT id, start_time, stop_time, note FROM stopwatches_entries WHERE stopwatch_id = ?  ORDER BY id DESC LIMIT ?`;
-  db.all(sql, [req.params.id, req.params.n], (err, rows) => {
+  const sql = `SELECT id, start_time, stop_time, note FROM stopwatches_entries WHERE stopwatch_id = ? AND user_id = ? ORDER BY id DESC LIMIT ?`;
+  db.all(sql, [req.params.id, req.session.id, req.params.n], (err, rows) => {
     if (err) {
       console.error(
         `Error getting entries for stopwatch ${req.params.id}:`,
@@ -41,8 +41,8 @@ router.get("/:id/entries/:n", (req, res) => {
 router.post("/:id/entries", async (req, res) => {
   const data = req.body;
   if (!(await stopwatchIsRunning(req.params.id))) {
-    const sql = `INSERT INTO stopwatches_entries(stopwatch_id, start_time) VALUES(?, ?)`;
-    const values = [req.params.id, data.start_time];
+    const sql = `INSERT INTO stopwatches_entries(stopwatch_id, user_id, start_time) VALUES(?, ?, ?)`;
+    const values = [req.params.id, req.session.id, data.start_time];
     db.run(sql, values, function (err) {
       if (err) {
         console.error(
@@ -72,8 +72,8 @@ router.post("/:id/entries", async (req, res) => {
 router.put("/:id/entries", async (req, res) => {
   const data = req.body;
   if (await stopwatchIsRunning(req.params.id)) {
-    const getEntryID = `SELECT id FROM stopwatches_entries WHERE stopwatch_id = ? ORDER BY id DESC LIMIT 1`;
-    const valuesForGetEntryID = [req.params.id];
+    const getEntryID = `SELECT id FROM stopwatches_entries WHERE stopwatch_id = ? AND user_id = ? ORDER BY id DESC LIMIT 1`;
+    const valuesForGetEntryID = [req.params.id, req.session.id];
     db.get(getEntryID, valuesForGetEntryID, (err, row) => {
       if (err) {
         console.error(
@@ -84,8 +84,8 @@ router.put("/:id/entries", async (req, res) => {
           error: `Error getting entry id for stopwatch ${req.params.id}`,
         });
       }
-      const sql = `UPDATE stopwatches_entries SET stop_time = ? WHERE id = ?`;
-      const values = [data.stop_time, row.id];
+      const sql = `UPDATE stopwatches_entries SET stop_time = ? WHERE id = ? AND user_id = ?`;
+      const values = [data.stop_time, row.id, req.session.id];
       db.run(sql, values, function (err) {
         if (err) {
           console.error(
@@ -114,8 +114,8 @@ router.put("/:id/entries", async (req, res) => {
 
 // Delete stopwatch entry of provided id
 router.delete("/entries/:entry_id", (req, res) => {
-  const sql = `DELETE FROM stopwatches_entries WHERE id = ?`;
-  db.run(sql, [req.params.entry_id], function (err) {
+  const sql = `DELETE FROM stopwatches_entries WHERE id = ? AND user_id = ?`;
+  db.run(sql, [req.params.entry_id, req.session.id], function (err) {
     if (err) {
       console.error(`Error deleting entry ${req.params.entry_id}:`, err);
       return res
@@ -129,8 +129,8 @@ router.delete("/entries/:entry_id", (req, res) => {
 // Edit note for a stopwatch entry of provided id
 router.put("/entries/:entry_id/note", (req, res) => {
   const data = req.body;
-  const sql = `UPDATE stopwatches_entries SET note = ? WHERE id = ?`;
-  const values = [data.note, req.params.entry_id];
+  const sql = `UPDATE stopwatches_entries SET note = ? WHERE id = ? AND user_id = ?`;
+  const values = [data.note, req.params.entry_id, req.session.id];
   db.run(sql, values, function (err) {
     if (err) {
       console.error(
